requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
data-importer==3.1.1
Django==3.1.7
django-channels==0.7.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==3.7.3
jdcal==1.4.1
kombu==5.0.2
oauthlib==3.1.0
openpyxl==2.4.0
prompt-toolkit==3.0.18
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.10.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
xlrd==1.0.0
zipp==3.4.1
